+++ app/src/test/java/com/android/purebilibili/MainActivityVideoRoutePolicyTest.kt
@@ -8,7 +8,7 @@ class MainActivityVideoRoutePolicyTest {
     @Test
     fun notificationRoute_defaultsToAutoPortraitAndStartAudioFalse() {
         assertEquals(
-            "video/BV1abc?cid=0&cover=&startAudio=false&autoPortrait=true&resumePositionMs=0",
+            "video/BV1abc?cid=0&cover=&startAudio=false&autoPortrait=true&resumePositionMs=0&commentRootRpid=0",
             resolveMainActivityVideoRoute(bvid = "BV1abc", cid = 0L)
         )
     }
@@ -16,7 +16,7 @@ class MainActivityVideoRoutePolicyTest {
     @Test
     fun notificationRoute_keepsCidValue() {
         assertEquals(
-            "video/BV2xyz?cid=12345&cover=&startAudio=false&autoPortrait=true&resumePositionMs=0",
+            "video/BV2xyz?cid=12345&cover=&startAudio=false&autoPortrait=true&resumePositionMs=0&commentRootRpid=0",
             resolveMainActivityVideoRoute(bvid = "BV2xyz", cid = 12345L)
         )
     }